Refactor the ResultSet class
Used std::vector<uint8_t> as the intermediate data type. Added required
overloads to mxs::Buffer::append to allow appending vectors.

Added a ResultSet::to_buffer method. This allows the resultset to be
written to the client using clientReply instead of writing it directly to
the DCB. The only user of the write function is schemarouter so
refactoring that should be easy.

Removed unused JSON conversion functions.