Add preliminary parser interface
The parser will call the Handler methods with the converted results. Most
types are given as strings, for the time being at least, as this reduces
the need for variant types in the base class.

Moved the response generation code in PinlokiSession into the handler
methods. Currently only placeholder values are returned.