Add initial parser implementation
The parser implementation implements most of the SQL required to start a
slave and configure replication.

Parsing for SHOW commands still needs to be implemented.

The whitespace handling in x3 is quite strange as both `STOP SLAVE` and
`STOPSLAVE` successfully parse as `STOP SLAVE`. This is because the parser
discards all whitespace between tokens. This can be fixed by adding
ignorable x3::ascii::space types between the keywords but that makes the
grammar a bit unclear and finding a solution that doesn't require this
would be better.