Fix log management commands
The PURGE BINARY LOGS syntax was wrong. The file up to which the binary
logs are purged is now passed to the handler. The BEFORE syntax for the
purging is not yet implemented.

Removed FLUSH LOGS as it would cause the binlog names and event positions
to divert from the ones on the master.