MXS-2922 Monitor according to CS version
The Columnstore monitor can monitor Columnstore version 1.0, 1.2
and 1.5. The 1.0 and 1.2 functionality is quite simplistic compared
to 1.5.

With 1.5 you cannot deduce much from being able to access the database
server, but must rely upon the information provided by the REST-API
daemon. This turns the startup it into something of a chicken/egg
problem as you must find out the version to use before you know what
needs to be done next. So in has_sufficient_permission() we start
by figuring out the version before checking whether we have the
actual permissions needed.

In 1.5 the functions mcsSystemReady(), mcsSystemReadOnly() and
mcsSystemPrimary() are no longer present.